Split oversized browser and comments views
